Fix field type for runtime to quiet MSVC warning

+++ src/client/ParseRequest.hpp
@@ -45,7 +45,7 @@ struct ParseRequest {
     int position = 0;
     int status = 0;
     long loc = 0;
-    long runtime = 0;
+    double runtime = 0;
     boost::optional<std::string> time_stamp;
     boost::optional<std::string> errormsg;
     bool needsparsing = true;
